ESP_NOW Practice
First round of learning ESP_NOW. The ESP receiver script takes an argument from the sender and outputs and analog and digital signal to two separate LEDs. The sender takes an analog input (from a pot) and a digital input (from a button) and sends the readings to the receiver via ESP_NOW. I'll add better comments later but for now this is working and I'm content with it.